* Extracted methods and enums of Toolbox into single purpose classes * Added AnimationClipOverrides * Added DiagionalDirection * Added ScaleLevel
