+++ Assets/Code/Classes/AnimationClipOverrides.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f89a863090f6f3d4ca76ecde506d2d02
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
