+++ Assets/Code/Classes/DiagonalDirection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a20f8451b9b16ef45beca41a6ddcdebb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
