+++ Assets/Code/Classes/Entities/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31be48ffca5a80243866c0486f418365
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
