+++ Assets/Code/Scripts/Entities/Player/ItemManager.cs
@@ -1,9 +1,5 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using static Toolbox;
-using static UnityEditor.Progress;
 
 public class ItemManager : MonoBehaviour
 {
