+++ Assets/Code/Scripts/Items/Weapons/Axe.cs
@@ -1,6 +1,5 @@
 using System.Linq;
 using UnityEngine;
-using static Toolbox;
 
 public class Axe : Weapon
 {
