+++ Assets/Code/Scripts/Items/Weapons/Greataxe.cs
@@ -1,8 +1,5 @@
-using System;
 using System.Linq;
-using TMPro;
 using UnityEngine;
-using static Toolbox;
 
 public class Greataxe : Weapon
 {
