+++ Assets/Code/Scripts/Items/Weapons/Longsword.cs
@@ -1,6 +1,5 @@
 using System.Linq;
 using UnityEngine;
-using static Toolbox;
 
 public class Longsword : Weapon
 {
